Fichier d'index fini et avancée sur la dichotomie

+++ Wikitionary/src/main/java/wiktionnary/wiktionnary/Coord.java
@@ -1,3 +1,3 @@
 package wiktionnary.wiktionnary;
 
-public record Coord(long debut, long fin){}
+public record Coord(int debut, int fin){}
